requirements.txt
django
djangorestframework
mozilla-django-oidc
africastalking